requirements.txt
flask
python-dotenv
redis